examples/sqla-pydantic/requirements.txt
starlette-admin
uvicorn
sqlalchemy
sqlalchemy-utils
pydantic[email]
